fix scrolltext bugs

+++ src/matrix.cpp
@@ -13,7 +13,7 @@ void Better_NeoMatrix::drawRGBPixel(int16_t x, int16_t y, uint8_t r, uint8_t g, 
   setPassThruColor();
 }
 
-void Better_NeoMatrix::scrollText(char* text, uint16_t length, uint8_t speed) {
+void Better_NeoMatrix::scrollText(char* text, uint16_t length, uint16_t speed) {
   uint32_t fullLengthInPixels = length * (dispCharWidth + 1) + displayWidth;
 
   for (uint32_t i = 0; i < fullLengthInPixels; i++) {
